requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.24.3
pyod==1.1.0
pandas==2.0.3

# PostgreSQL support (optional - only needed for production)
# psycopg2-binary==2.9.7
